+++ .github/workflows/test_filenames_check.yml
@@ -37,5 +37,6 @@ jobs:
           paths: '**/src/test/scala/**'
           report-format: 'console'
           excludes: '**/src/test/scala/za/co/absa/db/fadb/testing/**'
+                  slick/src/test/scala/za/co/absa/db/fadb/testng/classes/Actor.scala
           verbose-logging: 'false'
           fail-on-violation: 'true'
