+++ slick/src/test/scala/za/co/absa/db/fadb/testing/classes/Actor.scala
@@ -14,20 +14,6 @@
  * limitations under the License.
  */
 
-package za.co.absa.db.fadb.slick
+package za.co.absa.db.fadb.testing.classes
 
-import slick.jdbc.{GetResult, PositionedResult}
-
-/**
- *  A trait representing a converter from a Slick PositionedResult to an Actor.
- *  The trait is to be mixed into a SlickFunction returning an Actor.
- */
-trait ActorSlickConverter {
-
-  protected def slickConverter: GetResult[Actor] = {
-    def converter(r: PositionedResult): Actor = {
-      Actor(r.<<, r.<<, r.<<)
-    }
-    GetResult(converter)
-  }
-}
+case class Actor(actorId: Int, firstName: String, lastName: String)
