+++ slick/src/test/scala/za/co/absa/db/fadb/testing/classes/ActorSlickConverter.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.db.fadb.slick
+package za.co.absa.db.fadb.testing.classes
 
 import slick.jdbc.{GetResult, PositionedResult}
 
@@ -22,11 +22,11 @@ import slick.jdbc.{GetResult, PositionedResult}
  *  A trait representing a converter from a Slick PositionedResult to an Actor.
  *  The trait is to be mixed into a SlickFunction returning an Actor.
  */
-trait OptionalActorSlickConverter {
+trait ActorSlickConverter {
 
-  protected def slickConverter: GetResult[Option[Actor]] = {
-    def converter(r: PositionedResult): Option[Actor] = {
-      Some(Actor(r.<<, r.<<, r.<<))
+  protected def slickConverter: GetResult[Actor] = {
+    def converter(r: PositionedResult): Actor = {
+      Actor(r.<<, r.<<, r.<<)
     }
     GetResult(converter)
   }
